+++ src/vector/mod.rs
@@ -2,7 +2,6 @@
 
 pub mod aligned_buffer;
 pub mod background_compact;
-pub mod diskann;
 pub mod distance;
 pub mod f16;
 pub mod filter;